new way to handle translations
introduce `Tpl::setUiLanguage()` function to allow the application to
(re)set the language the templates should be rendered in.

The `Tpl` contructor no longer takes translation files, but folders
now. It will look for the specified UI language by finding all
supported languages and then checking if the provided UI language
is supported. It will silently fall back to `en-US`.